Create engine_communication and user_communication
